+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- As much as possible use subsections: Added, Removed, Modified, BugFix -->
 
+## [0.0.2] - 2023-08-26
+
+### Added
+
+- `--skip_verify`: Skip verification; only process files with no hash in filename
+- `--debug`: Debug mode (undocumented on purpose)
+
+### BugFix
+
+- fixed infinite recursion loop
+
 ## [0.0.1] - 2023-08-26
 
 __INITIAL RELEASE__